Move evaluator config mappings out of the replay config module
