+++ src/kitaru/server/adapters/rest/mapping/imports.py
@@ -18,11 +18,11 @@ from kitaru.api_models.v1.imports import (
     ImportListParams,
     ImportResponse,
 )
-from kitaru.server.adapters.rest.mapping.filtering import filter_to_expression
-from kitaru.server.adapters.rest.mapping.replay_config import (
+from kitaru.server.adapters.rest.mapping.evaluator_config import (
     evaluator_config_input,
     evaluator_config_to_wire,
 )
+from kitaru.server.adapters.rest.mapping.filtering import filter_to_expression
 from kitaru.server.application.models.imports import ImportCreate, ImportFilter
 from kitaru.server.domain.imports import Import
